feat(0X-Guide): first four object problems complete
