+++ 06-Objects/data/keyValidator.js
@@ -11,7 +11,18 @@ export const twoKeysTwo = {
   word: "hello",
   nums: [1, 2, 3],
 };
+
 export const twoKeysThree = {
   word: "banana",
   nums: [4, 5, 6],
+};
+
+export const threeKeysOne = {
+  phrase: "banana",
+  numbers: [4, 5, 6],
+};
+
+export const threeKeysOne = {
+  phrase: "banana",
+  numbers: [4, 5, 6],
 };
