+++ 06-Objects/data/objectCount.js
